requirements-model.txt
mlflow==2.21.0
kagglehub==0.3.13
pandas==2.3.3
python-dotenv==1.2.1
boto3